requirements.txt
pillow
rdkit